Refactoring the carousel filter as per Keith's comments
